Refactor: improve code structure and namespace usage
- Moved all public headers to the include/AmiClientCpp/ folder
- Wrapped all source/header files in `namespace ami {}` for better modular usage
- Replaced manual include guards with `#pragma once` for clarity
