+++ test/TestAmiClient.cpp
@@ -1,7 +1,8 @@
-﻿#include "../src/AmiClient.hpp"
-#include "../src/AmiClientListener.hpp"
-#include "../src/AmiTypes.hpp"
-#include "../src/AmiClientCommandDef.hpp"
+﻿#include <AmiClientCpp/AmiClient.hpp>
+#include <AmiClientCpp/AmiClientCommandDef.hpp>
+#include <AmiClientCpp/AmiClientListener.hpp>
+#include <AmiClientCpp/AmiTypes.hpp>
+
 #include <iostream>
 #include <memory>
 #include <thread>
@@ -9,94 +10,96 @@
 #include <mutex>
 #include <iomanip>
 
-extern std::mutex coutMutex;
+namespace ami {
 
-class MyAmiListener : public AmiClientListener {
-public:
-    void onConnect(AmiClient* client) override {
+    extern std::mutex coutMutex;
 
-        std::cout << "[Listener] Connected to server." << std::endl;
-    }
-
-
-    void onLoggedIn(AmiClient* client) override {
-
-        std::cout << "[Listener] Logged in successfully." << std::endl;
-        int ctr = 0;
-        while (ctr < 20) {
-            client->startObjectMessage("clienttest", "1")
-                .addMessageParamString("I", "bst_"+ std::to_string(ctr))
-                .addMessageParamString("name", "From_C++" + std::to_string(ctr))
-                .addMessageParamInt("age", ctr)
-                .sendMessageAndFlush();
-            ctr++;
-
-            std::this_thread::sleep_for(std::chrono::milliseconds(500));
+    class MyAmiListener : public AmiClientListener {
+    public:
+        void onConnect(AmiClient* client) override {
+            std::lock_guard<std::mutex> lk(coutMutex);
+            std::cout << "[Listener] Connected to server." << std::endl;
         }
 
+        void onLoggedIn(AmiClient* client) override {
+            {
+                std::lock_guard<std::mutex> lk(coutMutex);
+                std::cout << "[Listener] Logged in successfully." << std::endl;
+            }
 
+            // Send 20 test object messages
+            for (int ctr = 0; ctr < 20; ++ctr) {
+                client->startObjectMessage("clienttest", "1")
+                    .addMessageParamString("I", "bst_" + std::to_string(ctr))
+                    .addMessageParamString("name", "From_C++" + std::to_string(ctr))
+                    .addMessageParamInt("age", ctr)
+                    .sendMessageAndFlush();
 
-        AmiClientCommandDef def("sample_cmd_def");
-        def.setConditions({ AmiClientCommandDef::CONDITION_USER_CLICK })
-            .setName("ClickCommand")
-            .setHelp("Triggers on user click")
-            .setPriority(5);
+                std::this_thread::sleep_for(std::chrono::milliseconds(500));
+            }
 
-        client->sendCommandDefinition(def);
-    }
+            // Send a sample command definition
+            AmiClientCommandDef def("sample_cmd_def");
+            def.setConditions({ AmiClientCommandDef::CONDITION_USER_CLICK })
+                .setName("ClickCommand")
+                .setHelp("Triggers on user click")
+                .setPriority(5);
 
-    void onDisconnect(AmiClient* client) override {
+            client->sendCommandDefinition(def);
+        }
 
-        std::cout << "[Listener] Disconnected from server." << std::endl;
-    }
+        void onDisconnect(AmiClient* client) override {
+            std::lock_guard<std::mutex> lk(coutMutex);
+            std::cout << "[Listener] Disconnected from server." << std::endl;
+        }
 
-    void onMessageReceived(AmiClient* client,
-        long timestamp,
-        long seqNum,
-        int status,
-        const std::string& message) override {
+        void onMessageReceived(AmiClient* client,
+            long timestamp,
+            long seqNum,
+            int status,
+            const std::string& message) override {
+            std::lock_guard<std::mutex> lk(coutMutex);
+            std::cout << "[Listener] MessageReceived: seq=" << seqNum
+                << " status=" << status
+                << " msg=\"" << message << "\"" << std::endl;
+        }
 
- 
-        std::cout << "[Listener] MessageReceived: seq=" << seqNum
-            << " status=" << status
-            << " msg=\"" << message << "\"" << std::endl;
-    }
+        void onMessageSent(AmiClient* client,
+            const std::string& message) override {
+            std::lock_guard<std::mutex> lk(coutMutex);
+            std::cout << "[Listener] MessageSent: " << message << std::endl;
+        }
 
-    void onMessageSent(AmiClient* client,
-        const std::string& message) override {
+        void onCommand(AmiClient* source,
+            const std::string& requestId,
+            const std::string& cmd,
+            const std::string& userName,
+            const std::string& objectType,
+            const std::string& objectId,
+            const std::map<std::string, AmiValue>& params) override {
+            std::lock_guard<std::mutex> lk(coutMutex);
+            std::cout << "[Listener] Command received." << std::endl;
 
-        std::cout << "[Listener] MessageSent:" << message << std::endl;
-    }
+            source->startResponseMessage(requestId, 0, "Processed")
+                .addMessageParamLong("callback_code", 123)
+                .sendMessageAndFlush();
+        }
+    };
 
-    void MyAmiListener::onCommand(AmiClient* source,
-        const std::string& requestId,
-        const std::string& cmd,
-        const std::string& userName,
-        const std::string& objectType,
-        const std::string& objectId,
-        const std::map<std::string, AmiValue>& params)
-    {
-        std::cout << "[Listener] Command received:" << std::endl;
-
-        source
-            ->startResponseMessage(requestId, /*status=*/0, /*message=*/"Processed")
-            .addMessageParamLong("callback_code", 123)
-            .sendMessageAndFlush();
-    }
-
-};
+}  // namespace ami
 
 int main(int argc, char* argv[]) {
+    using namespace ami;
+
     std::string host = AmiClient::DEFAULT_HOST;
     int port = AmiClient::DEFAULT_PORT;
     std::string loginId = "demo";
-    bool autoFlush = false;
 
     if (argc > 1) host = argv[1];
     if (argc > 2) port = std::stoi(argv[2]);
     if (argc > 3) loginId = argv[3];
 
-    auto client = AmiClient::create(); 
+    auto client = AmiClient::create();
     auto listener = std::make_shared<MyAmiListener>();
     client->addListener(listener);
 
@@ -106,16 +109,12 @@ int main(int argc, char* argv[]) {
             << " with loginId=\"" << loginId << "\"..." << std::endl;
     }
 
-    int opts = AmiClient::ENABLE_AUTO_PROCESS_INCOMING ;
-    //int opts = AmiClient::ENABLE_AUTO_PROCESS_INCOMING | AmiClient::LOG_MESSAGES;
-    
+    int opts = AmiClient::ENABLE_AUTO_PROCESS_INCOMING;
 
     if (!client->start(host, port, loginId, opts)) {
         std::cerr << "[Main] Failed to start AmiClient." << std::endl;
         return 1;
     }
-
-    
 
     while (client->isConnected()) {
         std::this_thread::sleep_for(std::chrono::seconds(1));
