+++ test/TestMsgType.cpp
@@ -1,7 +1,7 @@
-﻿#include "../src/AmiClient.hpp"
-#include "../src/AmiClientListener.hpp"
-#include "../src/AmiTypes.hpp"
-#include "../src/AmiClientCommandDef.hpp"
+﻿#include <AmiClientCpp/AmiClient.hpp>
+#include <AmiClientCpp/AmiClientCommandDef.hpp>
+#include <AmiClientCpp/AmiClientListener.hpp>
+#include <AmiClientCpp/AmiTypes.hpp>
 #include <iostream>
 #include <memory>
 #include <thread>
@@ -9,77 +9,83 @@
 #include <mutex>
 #include <iomanip>
 
-extern std::mutex coutMutex;
+namespace ami {
+    extern std::mutex coutMutex;
 
-class MyAmiListener : public AmiClientListener {
-public:
-    void onConnect(AmiClient* client) override {
-        std::cout << "[Listener] Connected to server." << std::endl;
-    }
+    class MyAmiListener : public AmiClientListener {
+    public:
+        void onConnect(AmiClient* client) override {
+            std::cout << "[Listener] Connected to server." << std::endl;
+        }
 
-    void onLoggedIn(AmiClient* client) override {
-        std::cout << "[Listener] Logged in. Sending test messages..." << std::endl;
+        void onLoggedIn(AmiClient* client) override {
+            std::cout << "[Listener] Logged in. Sending test messages..." << std::endl;
 
-        // 1) Send an object creation message
-        
-        client->startObjectMessage("cmdtest", "test1")
-            .addMessageParamString("name", "jack")
-            .addMessageParamInt("number", 3)
-            .sendMessageAndFlush();
+            // 1) Send an object creation message
 
-        client->startObjectMessage("cmdtest", "test2")
-            .addMessageParamString("name", "mike")
-            .addMessageParamInt("number", 4)
-            .sendMessageAndFlush();
+            client->startObjectMessage("cmdtest", "test1")
+                .addMessageParamString("name", "jack")
+                .addMessageParamInt("number", 3)
+                .sendMessageAndFlush();
 
-        // 2) Send a delete message
-        std::this_thread::sleep_for(std::chrono::milliseconds(1000));
-        client->startDeleteMessage("cmdtest", "test1")
-            .sendMessageAndFlush();
+            client->startObjectMessage("cmdtest", "test2")
+                .addMessageParamString("name", "mike")
+                .addMessageParamInt("number", 4)
+                .sendMessageAndFlush();
 
-        // 3) Send a command definition
-        client->startCommandDefinition("bst")
-            .addMessageParamString("N", "2nd Bust Every Order")
-            .addMessageParamString("H", "busts all orders")
-            .addMessageParamInt("L", 2)
-            .sendMessageAndFlush();
-    }
+            // 2) Send a delete message
+            std::this_thread::sleep_for(std::chrono::milliseconds(1000));
+            client->startDeleteMessage("cmdtest", "test1")
+                .sendMessageAndFlush();
 
-    void onDisconnect(AmiClient* /*client*/) override {
-        std::cout << "[Listener] Disconnected from server." << std::endl;
-    }
+            // 3) Send a command definition
+            client->startCommandDefinition("bst")
+                .addMessageParamString("N", "2nd Bust Every Order")
+                .addMessageParamString("H", "busts all orders")
+                .addMessageParamInt("L", 2)
+                .sendMessageAndFlush();
+        }
 
-    void onMessageReceived(AmiClient* /*client*/,
-        long ts,
-        long seq,
-        int status,
-        const std::string& message) override {
-        std::cout << "[Listener] MessageReceived: seq=" << seq
-            << " status=" << status
-            << " msg=\"" << message << "\"" << std::endl;
-    }
+        void onDisconnect(AmiClient* /*client*/) override {
+            std::cout << "[Listener] Disconnected from server." << std::endl;
+        }
 
-    void onMessageSent(AmiClient* /*client*/, const std::string& msg) override {
-        std::cout << "[Listener] MessageSent: " << msg;
-    }
+        void onMessageReceived(AmiClient* /*client*/,
+            long ts,
+            long seq,
+            int status,
+            const std::string& message) override {
+            std::cout << "[Listener] MessageReceived: seq=" << seq
+                << " status=" << status
+                << " msg=\"" << message << "\"" << std::endl;
+        }
 
-    void onCommand(AmiClient* source,
-        const std::string& requestId,
-        const std::string& cmd,
-        const std::string& userName,
-        const std::string& objectType,
-        const std::string& objectId,
-        const std::map<std::string, AmiValue>& params) override {
-  
-        std::cout << "[Listener] Command received: id=" << requestId << " cmd=" << cmd << std::endl;
+        void onMessageSent(AmiClient* /*client*/, const std::string& msg) override {
+            std::cout << "[Listener] MessageSent: " << msg;
+        }
 
-        source->startResponseMessage(requestId, 0, "Processed")
-            .addMessageParamLong("callback_code", 123)
-            .sendMessageAndFlush();
-    }
-};
+        void onCommand(AmiClient* source,
+            const std::string& requestId,
+            const std::string& cmd,
+            const std::string& userName,
+            const std::string& objectType,
+            const std::string& objectId,
+            const std::map<std::string, AmiValue>& params) override {
+
+            std::cout << "[Listener] Command received: id=" << requestId << " cmd=" << cmd << std::endl;
+
+            source->startResponseMessage(requestId, 0, "Processed")
+                .addMessageParamLong("callback_code", 123)
+                .sendMessageAndFlush();
+        }
+    };
+
+    
+}
 
 int main(int argc, char* argv[]) {
+    using namespace ami;
+
     std::string host = AmiClient::DEFAULT_HOST;
     int port = AmiClient::DEFAULT_PORT;
     std::string loginId = "demo";
@@ -88,10 +94,9 @@ int main(int argc, char* argv[]) {
     if (argc > 2) port = std::stoi(argv[2]);
     if (argc > 3) loginId = argv[3];
 
-    std::string server_certificate_public_key_file = (argc > 4) ? argv[4] : "";
-    std::string client_certificate_public_key_file = (argc > 5) ? argv[5] : "";
-    std::string client_certificate_private_key_file = (argc > 6) ? argv[6] : "";
-
+    std::string serverCert = (argc > 4) ? argv[4] : "";
+    std::string clientPub = (argc > 5) ? argv[5] : "";
+    std::string clientKey = (argc > 6) ? argv[6] : "";
 
     auto client = AmiClient::create();
     auto listener = std::make_shared<MyAmiListener>();
@@ -103,24 +108,25 @@ int main(int argc, char* argv[]) {
             << " with loginId=\"" << loginId << "\"..." << std::endl;
     }
 
+    int opts = AmiClient::ENABLE_AUTO_PROCESS_INCOMING |
+        AmiClient::ENABLE_AUTO_FLUSH_OUTGOING |
+        AmiClient::ENABLE_SEND_SEQNUM |
+        AmiClient::ENABLE_SEND_TIMESTAMPS;
 
-    int opts = AmiClient::ENABLE_AUTO_PROCESS_INCOMING
-        | AmiClient::ENABLE_AUTO_FLUSH_OUTGOING
-        | AmiClient::ENABLE_SEND_SEQNUM
-        | AmiClient::ENABLE_SEND_TIMESTAMPS;
+    bool success = client->start(
+        host,
+        port,
+        loginId,
+        opts,
+        serverCert,
+        clientPub,
+        clientKey
+    );
 
-    if (!client->start(
-                host,
-                port,
-                loginId,
-                opts,
-                server_certificate_public_key_file,
-                client_certificate_public_key_file,
-                client_certificate_private_key_file)) {
+    if (!success) {
         std::cerr << "[Main] Failed to start AmiClient." << std::endl;
         return 1;
     }
-
 
     while (client->isConnected()) {
         std::this_thread::sleep_for(std::chrono::seconds(1));
